parent process exit status vs die_upon_destroy
With the debian packaged Proc::Background 1.08 and perl 5.10.0, the
program foo.pl below run as

    perl foo.pl; echo $?

gives an exit code of 1, where I hoped it would be 0 per the "exit 0".

I think most of the time it'd be a good thing if a Proc::Background
object destroy left the mainline's exit code in $? unchanged.  The only
concern would be whether anyone doing a die_upon_destroy in the middle
of a program did in fact want to see its status in "$?".  Though I'd
imagine if you want that you'd be doing ->die then ->wait, or whatever,
to get it.

fixes #2