Change handling of die(), and allow custom kill script
Previously, die() would call kill or TerminateProcess and then
a non-blocking waitpid and then sleep.  It is better to perform
a blocking waitpid for the desired duration (now that that is
supported by this module)

The sequence of kill signals was [HUP, QUIT, INT, KILL], each sent
five times.  That doesn't really match Unix expectations, where
TERM should be used for "please exit now" and KILL when forcing an
immediate exit.  Instead of hard-coding it, though, allow users to
specify a custom kill sequence.  Also, try to honor the TERM signal
for Windows without going directly to TerminateProcess, by calling
taskkill.exe.

fixes #1