Correct parsing/quoting of Win32 arguments
There were several bugs in the parsing and quoting of the command
line for Win32 processes.  Rather than digging deep into all the
details of this problem, just use the solution that already exists
in Win32::ShellQuote.

Note that according to the ShellQuote docs, there is no way to quote
arguments that parse the same from CommandLineToArgvW and by cmd.exe,
but this module doesn't call out to cmd.exe anyway so that shouldn't
be a problem.

fixes #7