Fix a bug related to alarm() on perl 5.8 & 5.10
For reasons unexplained, the Unix wait routine was failing to receive the
SIGALRM during waitpid.  An strace shows that alarm syscall was never called.
When using perl CORE::alarm() it works just fine.  This points to a bug in
Time::HiRes.  However, this bit of code was unchanged for a decade, so I
don't understand what related change caused it to break.  Regardless, perl
5.8 isn't all that important, so I just worked around it by using CORE::alarm
with the wait rounded up to a whole number of seconds.