Rename 'kill' to 'terminate'
Earlier I thought 'kill' was a better name than 'die', but since it
takes a list of signal names, Unix users might confuse it for a
general purpose signal-sender.  It only returns true if the process
exited, rather than whether a signal was sent, and multiple calls
to a dead process are not an error, so I renamed again to 'terminate'
to make the behavior clear.