Fix potential bug if system clock changes during timeout_system
I was checking for an early wake, and choosing the new delay based
on the system clock.  If the system clock were set backward, this
could wait a very long time.  I checked Time::HiRes's clock_gettime
but it is not available on Windows, so I made a less correct fix to
just never increase the delay.  So, a system clock change can grow
the available timeout but never by more than the original timeout
value.