Always detect errors up to exec() in Unix, via pipe trick
I added this feature conditional on 'autodie', but on further thought
I think everyone would want this improvement even if it slightly
changes the traditional semantics of how exec() errors are handled.
It isn't ever useful to think you've started a child process when one
of its hared libs was missing and immediately dies.